可以初始化已加密的版本库。

+++ Init.sh
@@ -12,11 +12,14 @@ echo ----------------------------------------
 echo           请输入本地库路径：
 echo ----------------------------------------
 read repoPath
-echo ----------------------------------------
-echo            请输入密码：
-echo ----------------------------------------
-read theKey
-cd $repoPath
+cd $repoPath 2>/dev/null
+if [ $? -eq 1 ]
+then
+    echo 版本库不存在，将创建新版本库。。。
+    mkdir $repoPath
+    cd $repoPath
+    git init
+fi
 repoPath=$(pwd)
 cd $mainpath
 reponame=$(echo $repoPath | grep -oP '(?<=/)[^/]*(?=$)')
@@ -24,19 +27,38 @@ if [ ! -d ~/.git_secure/$reponame ]
 then
     mkdir ~/.git_secure/$reponame
 fi
-cp *.template ~/.git_secure/$reponame
-cd ~/.git_secure/$reponame
-for template in $(ls *.template)
-do
-    chmod 777 $template
-    mv $template $(echo $template | grep -oP '.*(?=.template)')
-done
-sed -i 's/<your-password>/'$theKey'/g' *
-sed -i 's/<your-reponame>/'$reponame'/g' *
-touch hashandsalt
+if [ ! -f ~/.git_secure/$reponame/clean_filter_openssl ]
+then
+    existFlag=0
+    echo ----------------------------------------
+    echo            请输入密码：
+    echo ----------------------------------------
+    read theKey
+    cp *.template ~/.git_secure/$reponame
+    cd ~/.git_secure/$reponame
+    for template in $(ls *.template)
+    do
+	chmod 777 $template
+	mv $template $(echo $template | grep -oP '.*(?=.template)')
+    done
+    sed -i 's/<your-password>/'$theKey'/g' *
+    sed -i 's/<your-reponame>/'$reponame'/g' *
+    if [ ! -f hashandsalt ]
+    then
+	touch hashandsalt
+    fi
+else
+    existFlag=1
+    cd ~/.git_secure/$reponame
+fi
 cat ./config >> $repoPath/.git/config
 cat ./gitattributes >>$repoPath/.gitattributes
 echo ".gitattributes" >>$repoPath/.gitignore
+if [ existFlag -eq 1 ]
+then
+    cd $repoPath
+    git reset --hard HEAD
+fi
 echo ----------------------------------------
 echo               初始化完成。
 echo ----------------------------------------
